requirements.txt
eth-brownie==1.16.0
prettytable~=2.1.0
setuptools~=57.4.0